Se agrega lógica de owner y admin en el servicio

+++ services/comment.js
@@ -1,9 +1,9 @@
-const { Comment:DB } = require('../models')
+const { Comment: DB } = require('../models')
 const db = require('../models');
 
-const createComment = async( data ) => {
+const createComment = async (data) => {
 
-    const comment = new DB( data );
+    const comment = new DB(data);
 
     await comment.save();
 
@@ -12,21 +12,35 @@ const createComment = async( data ) => {
 };
 
 const findId = async (id) => {
-    
+
     const comment = await db.Comment.findByPk(id);
-    
-    return comment;
-
-}
-
-const deleteOne = async (id) => {
-
-    const comment = await db.Comment.destroy( {where: {id} });
 
     return comment;
 
 }
 
+const deleteOne = async (user, id) => {
+
+    const commentUserId = await db.Comment.findByPk(id);
+
+    if (commentUserId === null) {
+        return null;
+    } else {
+        console.log("Existe el comentario");
+        console.log("idComment: " + commentUserId.dataValues.user_id)
+        console.log("idUser: " + user.roleId)
+
+        if (commentUserId.dataValues.user_id === user.id || user.roleId === 1) {
+
+        const comment = await db.Comment.destroy({ where: { id } });
+
+        return comment;
+
+    } else {
+        return "No tiene permisos para eliminar este comentario";
+    }
+}
+}
 
 
 module.exports = {
